requirements/reqs-monitor-server.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==38.0.4
Django==4.1.4
django-admin==2.0.1
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.14.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-jsonview==2.0.0
django-six==1.0.4
django-widget-tweaks==1.4.12
exceptiongroup==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
multidict==6.0.3
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.6
pywin32==305
PyYAML==6.0
requests==2.28.1
screen==1.0.1
six==1.10.0
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
uvicorn==0.20.0
xlwt==1.3.0
yarl==1.8.2
python-dateutil~=2.8.2